gephi_export: Fix assigning colors to nodes
In some cases a node that is part of selected package wasn't assigned
a color because it was not in the custom list of nodes (all rules where
assigned to it via attribute).
Fix:
First add all edges to networkx graph and than assign colors to list
of nodes in the graph (instead of generating custom list of nodes as
before).